rename library to geometamaker
